+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # HidingBar
 
 * Fixed missing icon when updating Masque style.
+* Fixed OnEnter handler.